+++ src/components/User/Sales/components/SaleList.vue
@@ -159,14 +159,13 @@ export default {
 
   methods: {
     async getOwner () {
+      let owner
       const emailRegExp = /^[^\s@]+@[^\s@]+\.[^\s@]+$/
-      if (Sdk.base.Keypair.isValidPublicKey(this.owner)) {
-        this.filters.owner = this.owner
-      } else if (emailRegExp.test(this.owner)) {
-        this.filters.owner = await api.users.getAccountIdByEmail(this.owner)
-      } else {
-        this.filters.owner = this.owner
+      if (emailRegExp.test(this.owner)) {
+        owner = await api.users.getAccountIdByEmail(this.owner)
       }
+
+      return owner || this.owner
     },
 
     async getList () {
@@ -222,8 +221,8 @@ export default {
     'filters.openOnly' () {
       this.reloadCollectionLoader()
     },
-    'owner': _.throttle(function () {
-      this.getOwner()
+    'owner': _.throttle(async function () {
+      this.filters.owner = await this.getOwner()
       this.reloadCollectionLoader()
     }, 1000),
     'filters.name': _.throttle(function () {